Switch snapshot configuration extensions to a custom version
