+++ ElementX/Sources/Screens/EmojiPickerScreen/View/EmojiPickerScreen.swift
@@ -95,7 +95,7 @@ struct EmojiPickerScreen_Previews: PreviewProvider, TestablePreview {
     static var previews: some View {
         EmojiPickerScreen(context: viewModel.context, selectedEmojis: ["😀", "😄"])
             .previewDisplayName("Screen")
-            .snapshot(delay: 0.5)
+            .snapshotPreferences(delay: 0.5)
     }
 }
