+++ ElementX/Sources/Screens/RoomDetailsScreen/View/RoomDetailsScreen.swift
@@ -401,12 +401,12 @@ struct RoomDetailsScreen_Previews: PreviewProvider, TestablePreview {
     static var previews: some View {
         RoomDetailsScreen(context: simpleRoomViewModel.context)
             .previewDisplayName("Simple Room")
-            .snapshot(delay: 2)
+            .snapshotPreferences(delay: 2)
         RoomDetailsScreen(context: dmRoomViewModel.context)
             .previewDisplayName("DM Room")
-            .snapshot(delay: 0.25)
+            .snapshotPreferences(delay: 0.25)
         RoomDetailsScreen(context: genericRoomViewModel.context)
             .previewDisplayName("Generic Room")
-            .snapshot(delay: 0.25)
+            .snapshotPreferences(delay: 0.25)
     }
 }
