+++ ElementX/Sources/Screens/RoomDirectorySearchScreen/View/RoomDirectorySearchScreen.swift
@@ -107,6 +107,6 @@ struct RoomDirectorySearchScreen_Previews: PreviewProvider, TestablePreview {
     
     static var previews: some View {
         RoomDirectorySearchScreen(context: viewModel.context)
-            .snapshot(delay: 1.0)
+            .snapshotPreferences(delay: 1.0)
     }
 }
