+++ ElementX/Sources/Screens/RoomMemberListScreen/View/RoomMembersListManageMemberSheet.swift
@@ -81,13 +81,13 @@ struct RoomMembersListManageMemberSheet_Previews: PreviewProvider, TestablePrevi
                                          actions: [.kick, .ban],
                                          context: viewModel.context)
             .previewDisplayName("Joined")
-            .snapshot(delay: 0.2)
+            .snapshotPreferences(delay: 0.2)
         
         RoomMembersListManageMemberSheet(member: .init(withProxy: RoomMemberProxyMock.mockBanned[3]),
                                          actions: [],
                                          context: viewModel.context)
             .previewDisplayName("Banned")
-            .snapshot(delay: 0.2)
+            .snapshotPreferences(delay: 0.2)
     }
 }
