+++ ElementX/Sources/Screens/RoomPollsHistoryScreen/View/RoomPollsHistoryScreen.swift
@@ -160,12 +160,12 @@ struct RoomPollsHistoryScreen_Previews: PreviewProvider, TestablePreview {
             RoomPollsHistoryScreen(context: viewModelEmpty.context)
         }
         .previewDisplayName("No polls")
-        .snapshot(delay: 1.0)
+        .snapshotPreferences(delay: 1.0)
 
         NavigationStack {
             RoomPollsHistoryScreen(context: viewModel.context)
         }
         .previewDisplayName("polls")
-        .snapshot(delay: 1.0)
+        .snapshotPreferences(delay: 1.0)
     }
 }
