+++ ElementX/Sources/Screens/SecureBackup/SecureBackupScreen/View/SecureBackupScreen.swift
@@ -157,7 +157,7 @@ struct SecureBackupScreen_Previews: PreviewProvider, TestablePreview {
             }
             .previewDisplayName("Recovery incomplete")
         }
-        .snapshot(delay: 1.0)
+        .snapshotPreferences(delay: 1.0)
     }
     
     static func viewModel(keyBackupState: SecureBackupKeyBackupState,
