+++ ElementX/Sources/Screens/Timeline/View/Replies/TimelineReplyView.swift
@@ -326,7 +326,7 @@ struct TimelineReplyView_Previews: PreviewProvider, TestablePreview {
         .padding()
         .environmentObject(viewModel.context)
         // Allow member names to load. Reduce precission as the `imageSource` randomly renders slightly differently
-        .snapshot(delay: 0.2, precision: 0.98)
+        .snapshotPreferences(delay: 0.2, precision: 0.98)
         .previewLayout(.sizeThatFits)
     }
 }
