+++ ElementX/Sources/Screens/Timeline/View/TimelineItemViews/FormattedBodyText.swift
@@ -214,7 +214,7 @@ struct FormattedBodyText_Previews: PreviewProvider, TestablePreview {
             .padding()
         }
         .previewLayout(.sizeThatFits)
-        .snapshot(delay: 0.25)
+        .snapshotPreferences(delay: 0.25)
     }
 }
